+++ vast-knowledge-service/vast-knowledge-service-article/src/main/java/com/vk/article/domain/ApCollection.java
@@ -9,7 +9,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 
 /**
@@ -26,17 +26,17 @@ import java.time.LocalDateTime;
 public class ApCollection implements Serializable {
 
     @Id(keyType = KeyType.Auto)
-    private BigInteger id;
+    private Long id;
 
     /**
      * 实体ID
      */
-    private BigInteger entryId;
+    private Long entryId;
 
     /**
      * 文章ID
      */
-    private BigInteger articleId;
+    private Long articleId;
 
     /**
      * 点赞内容类型 0:文章 1:动态
