+++ vast-knowledge-service/vast-knowledge-service-user/src/main/java/com/vk/user/domain/ApUserSearch.java
@@ -9,7 +9,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 
 /**
@@ -29,12 +29,12 @@ public class ApUserSearch implements Serializable {
      * 主键
      */
     @Id(keyType = KeyType.Auto)
-    private BigInteger id;
+    private Long id;
 
     /**
      * 用户ID
      */
-    private BigInteger entryId;
+    private Long entryId;
 
     /**
      * 搜索词
